refactor: no components definitions inside another
move EndpointManager from hooks to class folder
