+++ packages/extension-polkagate/src/fullscreen/governance/FullScreenHeader.tsx
@@ -10,7 +10,7 @@ import { useParams } from 'react-router';
 import { logoBlack } from '../../assets/logos';
 import { ActionContext } from '../../components';
 import { useApi, useChain, useGenesisHashOptions } from '../../hooks';
-import { FullScreenChainSwitch, FullScreenRemoteNode } from '../../partials';
+import { FullScreenChainSwitch, RemoteNodeSelectorWithSignals } from '../../partials';
 import { EXTENSION_NAME, GOVERNANCE_CHAINS, IDENTITY_CHAINS, SOCIAL_RECOVERY_CHAINS, STAKING_CHAINS } from '../../util/constants';
 import { openOrFocusTab } from '../accountDetails/components/CommonTasks';
 import AddressDropdown from './components/AddressDropdown';
@@ -114,7 +114,7 @@ export function FullScreenHeader ({ _otherComponents, noAccountDropDown = false,
                 </Grid>
                 {chain &&
                   <Grid container item justifyContent='flex-end' width='50px'>
-                    <FullScreenRemoteNode
+                    <RemoteNodeSelectorWithSignals
                       address={address}
                     />
                   </Grid>
