+++ packages/extension-polkagate/src/partials/index.ts
@@ -1,18 +1,15 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
-export { default as HeaderBrand } from './HeaderBrand';
-// export { default as MenuSettings } from './MenuSettings';
-export { default as Name } from './Name';
-export { default as ChainSwitch } from './ChainSwitch';
-export { default as Passwords } from './Passwords';
 export { default as AccountMenu } from './AccountMenu';
-// export { default as AccMenuInside } from './AccMenuInside';
-export { default as WaitScreen } from './WaitScreen';
-export { default as SubTitle } from './SubTitle';
-export { default as ThroughProxy } from './ThroughProxy';
+export { default as AddNewAccountButton } from './AddNewAccountButton';
+export { default as ChainSwitch } from './ChainSwitch';
 export { default as Confirmation } from './Confirmation';
 export { default as FullScreenChainSwitch } from './FullScreenChainSwitch';
-export { default as FullScreenRemoteNode } from './FullScreenRemoteNode';
-export { default as AddNewAccountButton } from './AddNewAccountButton';
+export { default as HeaderBrand } from './HeaderBrand';
+export { default as Name } from './Name';
+export { default as Passwords } from './Passwords';
+export { default as RemoteNodeSelectorWithSignals } from './RemoteNodeSelectorWithSignals';
+export { default as SubTitle } from './SubTitle';
+export { default as ThroughProxy } from './ThroughProxy';
+export { default as WaitScreen } from './WaitScreen';
